+++ CHANGELOG.md
@@ -34,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Increase frequency resolution in `set_periodic` (#686)
 - Fixed ESP32-S3 radio clock gating (#679)
 - Fixed ESP32-S2 radio clock gating (#681)
+- Fixed ESP32 and ESP32-S2 radio clock gating (#681)
+- Fixed ESP32, ESP32-S2, ESP32-C2 radio clock gating (#681)
 
 ### Removed